+++ tests/Frank.Tests/Program.fs
@@ -1,5 +1,0 @@
-module Program
-
-[<EntryPoint>]
-let main args =
-    0